Handle realm list command

+++ src/main/groovy/com/woodlands/yanp/auth/db/repository/AccountBannedRepository.groovy
@@ -24,5 +24,5 @@ interface AccountBannedRepository extends JpaRepository<AccountBannedEntity, Lon
     // TODO CMangos only pulls the banned_at and expires_at columns - is that actually more performant? Enough that we should we replicate that behavior?
     // It's a lot trickier to do cleanly in JPA https://www.w3docs.com/snippets/java/spring-jpa-selecting-specific-columns.html
 
-    AccountBannedEntity findByIdAndIsActive(Long id, boolean active)
+    AccountBannedEntity findByIdAndIsActive(Integer id, boolean active)
 }
